+++ deps/exokit-bindings/webrtc/src/getusermedia.cc
@@ -15,7 +15,7 @@
 #include "src/converters/dictionaries.h"  // IWYU pragma: keep
 #include "src/converters/interfaces.h"  // IWYU pragma: keep
 #include "src/converters/object.h"  // IWYU pragma: keep
-#include "src/converters/v8.h"  // IWYU pragma: keep
+#include "src/converters/v8-converters.h"  // IWYU pragma: keep
 #include "src/error.h"
 #include "src/events.h"  // IWYU pragma: keep
 #include "src/functional/curry.h"  // IWYU pragma: keep
